requirements.txt
# Core dependencies
tensorflow>=2.10.0
numpy==1.26.4
pillow>=9.0.0
matplotlib>=3.5.0

# Web frameworks
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
streamlit==1.21.0

# Image processing and GIS
rasterio>=1.3.0
geopandas>=0.12.0
shapely>=2.0.0
scikit-image>=0.19.0

# Data handling
pandas>=1.5.0
scipy>=1.9.0

# Utilities
python-multipart>=0.0.6
requests>=2.28.0
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0

# Visualization
seaborn>=0.12.0
plotly>=5.13.0
